+++ serde_arrow/src/internal/utils/array_ext.rs
@@ -3,7 +3,7 @@
 use marrow::array::{BytesArray, BytesViewArray, PrimitiveArray};
 
 use crate::internal::{
-    error::{fail, Result},
+    error::{fail, Error, ErrorKind, Result},
     utils::Offset,
 };
 
@@ -499,7 +499,10 @@ pub fn set_validity(buffer: Option<&mut Vec<u8>>, idx: usize, value: bool) -> Re
     } else if value {
         Ok(())
     } else {
-        fail!("Cannot push null for non-nullable array");
+        Err(Error::new(
+            ErrorKind::NullabilityViolation { field: None },
+            "Cannot push null for non-nullable array".into(),
+        ))
     }
 }