+++ serde_arrow/src/test/error_messages/misc.rs
@@ -1,9 +1,9 @@
-use crate::internal::error::Error;
+use crate::internal::error::{Error, ErrorKind};
 
 #[test]
 #[allow(clippy::bool_assert_comparison)]
 fn backtrace_on_debug() {
-    let err = Error::custom(String::from("foo bar"));
+    let err = Error::new(ErrorKind::Custom, String::from("foo bar"));
 
     // NOTE: the exact message depends on the ability of Rust to capture a backtrace
     assert!(!format!("{}", err).contains("Backtrace"));